pyproject.toml
[tool.kedro]
package_name = "llms-rag-enx"
project_name = "llms-rag-enx"

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
addopts = """
--cov-report term-missing \
--cov src/dpt -ra"""

[tool.coverage.report]
fail_under = 0
show_missing = true
exclude_lines = ["pragma: no cover", "raise NotImplementedError"]

[tool.poetry]
name = "llms-rag-enx"
version = "0.0.1"
description = ""
authors = ["Duc Thang Hoang <ducthang.hoang@euranova.eu>", "Yanis Labeyrie <yanis.labeyrie@euranova.eu>"]
readme = "README.md"
repository = "https://code.euranova.eu/demolabo/llms-and-rag"

[tool.poetry.dependencies]
python = "^3.10, <3.11"
torch = "^2.2.1"
datasets = "2.14.0"
groq = "^0.5.0"
keybert = "0.8.4"

langchain = "<0.3.0"
langchain-community = "<0.3.0"  
langchain-groq = "<0.3.0"  
langchain-core = "<0.3.0" 
langchain-cerebras = "^0.1.0"
langchain-experimental = "<0.3.0" 
langchain-unstructured = {extras = ["local"], version = "^0.1.2"}


langsmith = "^0.1.84"
matplotlib = "3.8.0"
mattermostdriver = "7.3.2"
numpy = "1.26.4"
pandas = "2.2.2"
pydantic = "^2.0"
python-dotenv = "1.0.1"
pyyaml = "6.0.1"
requests = "^2.32.2"
rouge = "1.0.1"
scikit-learn = "1.3.0"
selenium = "4.19.0"
sentence-transformers = "3.0.1"
st-annotated-text = "4.0.1"
streamlit = "^1.35.0"
tqdm = "^4.66.4"
umap-learn = "0.5.6"
fitz = "0.0.1.dev2"
docx2txt = "0.8"
openpyxl = "3.1.2"
notebook = "7.1.3"
ruff = "0.4.1"
loguru = "^0.7.2"
cryptography = "^42.0.7"
ollama = "0.3.3"
thefuzz = "^0.22.1"
pydub = "^0.25.1"
moviepy = "^1.0.3"
llmlingua = "^0.2.2"
accelerate = "^0.30.1"
python-gitlab = "^4.5.0"
soundfile = "^0.12.1"
faster-whisper = "^1.0.2"
bayesian-optimization = "^1.4.3"
wordfreq = "^3.1.1"
qrcode = "^7.4.2"
markdown = "^3.6"
streamlit-audiorec = "^0.1.3"
streamlit-audiorecorder = "^0.0.5"
rank-bm25 = "^0.2.2"
setfit = "^1.0.3"
huggingface = "^0.0.1"
transformers = "^4.41.2"
insanely-fast-whisper = "^0.0.15"
packaging = "<24.0"
wheel = "^0.43.0"
fpdf = "^1.7.2"
bertopic = "^0.16.2"
translate = "^3.6.1"
deep-translator = "^1.11.4"
gliner = "^0.2.2"
gliner-spacy = "^0.0.8"
langchain-qdrant = "^0.1.0"
seaborn = "^0.13.2"
langchain-huggingface = "^0.0.3"
text-generation = "^0.7.0"
google-search-results = "^2.4.2"
numexpr = "2.8.4"
langchainhub = "^0.1.20"
sentencepiece = "^0.2.0"
jinja2 = "^3.1.4"
vocos = "^0.1.0"
einops = "^0.8.0"
streamlit-pills = "^0.3.0"
pdfplumber = "^0.11.1"
pymupdf = "^1.24.7"
pdf2image = "^1.17.0"
timm = "^1.0.7"
python-doctr = "^0.8.1"
jq = "^1.7.0"
glirel = "^0.1.4"
cupy-cuda12x = "^13.2.0"
flagembedding = "^1.2.10"
grand-cypher = "^0.9.0"
huggingface-hub = "0.23.5"
imageio = "2.34.1"
python-igraph = "^0.11.6"
leidenalg = "^0.10.2"
python-bidi = "0.4.2"
easyocr = "^1.7.1"
pipdeptree = "^2.23.1"
enum34 = "^1.1.10"
pytorch-metric-learning = "2.6.0"
peft = "^0.12.0"
diffusers = {git = "https://github.com/huggingface/diffusers.git"}
mlflow = "^2.15.1"
pytube = "^15.0.0"
youtube-dl = "^2021.12.17"
pytubefix = "^6.9.2"
spicy = "^0.16.0"
notion-client = "^2.2.1"
notion = "^0.0.28"
spacy = "^3.7.5"
sphinx = ">=5,<8"
sphinx-rtd-theme = "^2.0.0"
furo = "^2024.8.6"
sphinx-press-theme = "^0.9.1"
pyinstaller = "^6.10.0"
streamlit-lottie = "^0.0.5"
zxing-cpp = "^2.2.0"
texify = "^0.1.10"
torchao = "^0.4.0"
qwen-vl-utils = "^0.0.4"
st-multimodal-chatinput = "^0.2.1"
lingua-language-detector = "^2.0.2"
gradio-client = "^1.3.0"
aiofiles = "^24.1.0"
python-pptx = "^1.0.1"
webdriver-manager = "^4.0.2"
prometheus-client = "^0.20.0"

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    ".cache",
    ".local",
    "__pypackages__",
    "*.pyi",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]
# Same as Black.
line-length = 80
indent-width = 4
